temp

+++ include/Feature.h
@@ -26,4 +26,4 @@ namespace mvo
 }//namespace mvo
 
 void ManageTrackPoints(const mvo::Feature& present, mvo::Feature& before);
-void ManageTrackPoints(const mvo::Feature& present, mvo::Feature& before, std::vector<cv::Point3d>& mapPoints);
+void ManageTrackPoints(const mvo::Feature& present, mvo::Feature& before, std::vector<cv::Point3f>& mapPoints);
